refactor(voice/capture): extract LifecycleMixin (stream open/close/shutdown) [T1.4 step 9a]
Extract the three stream-lifecycle helpers from
``voice/_capture_task.py`` into a new ``LifecycleMixin`` class at
``voice/capture/_lifecycle_mixin.py`` per CLAUDE.md anti-pattern
#16:

* ``_signal_consumer_shutdown`` (~23 LOC) — flips ``_running`` +
  cancels the consumer task; called by the terminal
  ``OPEN_FAILED_NO_STREAM`` branches of ``request_exclusive_restart``
  and ``request_shared_restart`` (both already on RestartMixin).
* ``_emit_stream_opened`` (~47 LOC) — generates a fresh
  ``stream_id``, resets per-stream xrun + sustained-underrun
  counters, and emits ``audio.stream.opened``.
* ``_close_stream`` (~32 LOC) — emits ``audio.stream.closed``
  with cumulative xrun counts before tearing the stream down;
  tolerant of already-closed streams.

Total: ~105 LOC migrated. ``_capture_task.py`` reduces from 1170
to 1065 LOC.

The mixin declares 13 host-class state attributes (``_running``,
``_consumer``, ``_stream``, ``_stream_id``, the three xrun
counters, the four sustained-underrun window snapshots,
``_blocksize``, ``_resolved_device_name``) for mypy strict; all
three methods read/write that state directly so no method-via-MRO
hooks are required on the mixin itself.

Class composition:
  AudioCaptureTask(EpochMixin, RingMixin, RestartMixin)
    -> AudioCaptureTask(EpochMixin, RingMixin, LifecycleMixin, RestartMixin)

LifecycleMixin sits BEFORE RestartMixin in the MRO so the latter's
method-via-MRO declarations for ``_close_stream``,
``_emit_stream_opened``, and ``_signal_consumer_shutdown`` resolve
to the new mixin (was: AudioCaptureTask). MRO order otherwise
unchanged.

Test impact: zero — the three lifecycle methods have no
``patch("sovyx.voice._capture_task.<method>", …)`` test sites
(verified via grep). One docstring reference in
``test_capture_task.py`` remains and is correct as a documentation
mention.

Step 9b will land the audio-thread + consume-loop machinery
(``_audio_callback``, ``_enqueue``, ``_check_sustained_underrun_rate``,
``_consume_loop``, ``_maybe_emit_heartbeat``) in a follow-up
commit. That step requires migrating ~10
``patch("sovyx.voice._capture_task.time.monotonic", …)`` test
sites to follow the split (CLAUDE.md anti-pattern #20) so it gets
its own commit boundary.

Gates: ruff + ruff format + mypy strict (428 files) + bandit
clean. Full pytest 11894 passed (zero delta vs step 8b).

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 step 9a
